add pagination

+++ public/expenses.php
@@ -1,10 +1,9 @@
 <?php require_once("../includes/initialize.php")?>
 <?php
     global $database;
-    
-    $expenses_set = Expense::get_all_month_expenses();
+
     // $expenses = Expense::forMonth('June')->all();
-    $number_of_expenses = $database->num_rows($expenses_set)  ;
+    $number_of_expenses = $database->num_rows(Expense::get_all_month_expenses())  ;
     $number_of_expenses_per_page = 6 ;
     $number_of_pages= ceil((float)$number_of_expenses/(float)$number_of_expenses_per_page);
 
@@ -14,6 +13,17 @@
             redirect_to("not_available.php");
         }
     }
+
+    $pagination = new Pagination($page_number,$number_of_expenses_per_page,$number_of_expenses);
+
+    $sql = "SELECT * FROM expenses " ;
+    $sql .= "WHERE user_id = ".$_SESSION["user_id"] ;
+    $sql .= " ORDER BY id DESC ";
+    $sql .= " LIMIT ".$pagination->per_page ;
+    $sql .= " OFFSET ".$pagination->offset() ;
+    
+    $expenses_set  = $database->query($sql);
+
 ?>
 
 
@@ -39,15 +49,9 @@
             <?php
                 
                 $iteration_number = 0 ;
-                $iteration_number_to_escape = 0 ;
-
                 if($expenses_set != null){ 
                     while($expense=mysqli_fetch_assoc($expenses_set)){
 
-                    if($iteration_number_to_escape < ( ($page_number - 1) *$number_of_expenses_per_page)){
-                        ++$iteration_number_to_escape ;
-                        continue ;
-                    }
                     if($iteration_number == $number_of_expenses_per_page){
                         break ;
                     }else{
@@ -60,7 +64,7 @@
 
                         <td><?php echo $expense["expense_name"] ?></td>
                         <td><?php echo $expense["price"] ?></td>
-                        <td><?php echo $expense["category"] ?></td>
+                        <td><?php echo ucfirst($expense["category"]) ?></td>
                         <td><?php echo $expense["comment"] ?></td>
                         <td><?php echo $expense["created_at"] ?></td>
                         
@@ -68,7 +72,7 @@
                             <div class="btn-action">
                                     <a class= "btn-action-edit" href="edit_expense.php?expenseid=<?php echo $expense["id"] ?>"  value="edit">
                                             <img src="images/edit.png" class="btn-action-edit-image" alt="edit"></a>
-                                    <a class= "btn-action-delete" href="delete_expense.php?expenseid=<?php echo $expense["id"] ?>"  value="delete">
+                                    <a class= "btn-action-delete" href="delete_expense.php?expenseid=<?php echo $expense["id"] ?>"  value="delete" onclick="return confirm('Are you sure?');">
                                         <img src="images/delete.png" class="btn-action-delete-image" alt="delete"></a>
                                     </a>
                             </div>
@@ -90,13 +94,13 @@
 
     <?php
         echo "<a";
-        if($page_number > 1){
+        if($pagination->has_prev_page()){
             echo " href=\"" ;
             echo " ?pagenumber=" ;
-            echo  $page_number-1  ;
+            echo  "{$pagination->prev_page()}"  ;
             echo "\"" ; 
         }else{
-            echo " href=\"\"" ;
+            echo " href=\"?pagenumber={$pagination->current_page()}\"" ;
         }
         echo "class=\"btn-list-back btn\"" ;
         echo ">";
@@ -105,18 +109,27 @@
     ?>
 
     <span class="btn-list-page_number">
-        <?php echo $page_number?>
+        <?php 
+            for($i=1;$i <= $pagination->total_pages();$i++){ 
+                if($i == $pagination->current_page()){
+                    echo "<span class=\"btn-list-page_number-selected\">{$i}</span>" ;
+                }else{
+                    echo "<a href=\"?pagenumber={$i}\"  class=\"btn-list-page_number-link\">{$i}</a>" ;
+                }
+            }
+        ?>
     </span>
+    
     
     <?php
         echo "<a";
-        if($page_number < ($number_of_pages)){
+        if($pagination->has_next_page()){
             echo " href=\"" ;
             echo " ?pagenumber=" ;
-            echo  $page_number+1  ;
+            echo  "{$pagination->next_page()}"  ;
             echo "\"" ; 
         }else{
-            echo " href=\"\"" ;
+            echo " href=\"?pagenumber={$pagination->current_page()}\"" ;
         }
         echo "class=\"btn-list-next btn\"" ;
         echo ">";
@@ -125,5 +138,5 @@
     ?>
 
 </div>
-<?php mysqli_free_result($expenses_set); ?>
+<?php $database->free_result($expenses_set); ?>
 <?php include("layouts/footer.php")?>
