+++ public/search.php
@@ -23,11 +23,19 @@
     $number_of_pages= ceil((float)$number_of_expenses/(float)$number_of_expenses_per_page);
 
     $page_number = get_page_number() ;
-    if(($page_number > $number_of_pages) || ($page_number < 1)){
-        if($number_of_pages != 0){ 
-            redirect_to("not_available.php");
-        }
+    
+        
+    
+    if(($number_of_expenses == 0)){
+        $_SESSION["message"] = "No Matching" ;
+        redirect_to("index.php");
     }
+    elseif(($page_number > $number_of_pages) || ($page_number < 1)){
+        redirect_to("not_available.php");
+    }
+
+
+    $pagination = new Pagination($page_number,$number_of_expenses_per_page,$number_of_expenses);
 
 ?>
 
@@ -74,7 +82,7 @@
 
                         <td><?php echo $expense["expense_name"] ?></td>
                         <td><?php echo $expense["price"] ?></td>
-                        <td><?php echo $expense["category"] ?></td>
+                        <td><?php echo ucfirst($expense["category"]) ?></td>
                         <td><?php echo $expense["comment"] ?></td>
                         <td><?php echo $expense["created_at"] ?></td>
                         
@@ -82,7 +90,7 @@
                             <div class="btn-action">
                                     <a class= "btn-action-edit" href="edit_expense.php?expenseid=<?php echo $expense["id"] ?>"  value="edit">
                                             <img src="images/edit.png" class="btn-action-edit-image" alt="edit"></a>
-                                    <a class= "btn-action-delete" href="delete_expense.php?expenseid=<?php echo $expense["id"] ?>"  value="delete">
+                                            <a class= "btn-action-delete" href="delete_expense.php?expenseid=<?php echo $expense["id"] ?>"  value="delete" onclick="return confirm('Are you sure?');">
                                         <img src="images/delete.png" class="btn-action-delete-image" alt="delete"></a>
                                     </a>
                             </div>
@@ -104,13 +112,13 @@
 
     <?php
         echo "<a";
-        if($page_number > 1){
+        if($pagination->has_prev_page()){
             echo " href=\"" ;
-            echo " ?searchfor=$search_string&pagenumber=" ;
-            echo  $page_number-1  ;
+            echo " ?searchfor={$search_string}&pagenumber=" ;
+            echo  "{$pagination->prev_page()} "  ;
             echo "\"" ; 
         }else{
-            echo " href=\"\"" ;
+            echo " href=\"?searchfor={$search_string}&pagenumber={$pagination->current_page()}\"" ;
         }
         echo "class=\"btn-list-back btn\"" ;
         echo ">";
@@ -118,19 +126,28 @@
         echo "</a> " ;
     ?>
 
+
     <span class="btn-list-page_number">
-        <?php echo $page_number?>
+        <?php 
+            for($i=1;$i <= $pagination->total_pages();$i++){ 
+                if($i == $pagination->current_page()){
+                    echo "<span class=\"btn-list-page_number-selected\">{$i}</span>" ;
+                }else{
+                    echo "<a href=\"?searchfor={$search_string}&pagenumber={$i}\"  class=\"btn-list-page_number-link\">{$i}</a>" ;
+                }
+            }
+        ?>
     </span>
     
     <?php
         echo "<a";
-        if($page_number < ($number_of_pages)){
+        if($pagination->has_next_page()){
             echo " href=\"" ;
-            echo " ?searchfor=$search_string&pagenumber=" ;
-            echo  $page_number+1  ;
+            echo " ?searchfor={$search_string}&pagenumber=" ;
+            echo  "{$pagination->next_page()} "  ;
             echo "\"" ; 
         }else{
-            echo " href=\"\"" ;
+            echo " href=\"?searchfor={$search_string}&pagenumber={$pagination->current_page()}\"" ;
         }
         echo "class=\"btn-list-next btn\"" ;
         echo ">";
@@ -140,5 +157,5 @@
 
 </div>
 
-<?php mysqli_free_result($expenses_set); ?>
+<?php $database->free_result($expenses_set); ?>
 <?php include("layouts/footer.php")?>
